requirements.txt
Flask
flask_sqlalchemy
flask_migrate
python-dotenv
sklearn
celery
psycopg2-binary